lancamento programado finish

+++ web/app/js/controller/financeiro/lancamento/lancamentoController.js
@@ -100,6 +100,7 @@ app.controller('LancamentoController', ['$scope', '$filter', 'LancamentoService'
             return _.map(data.lancamentos, function(lancamento) { 
                 if(lancamento.tipo.codigo === 'despesa') { lancamento.pagamento = lancamento.valor * -1; } 
                 else if(lancamento.tipo.codigo === 'receita') { lancamento.deposito = lancamento.valor; }  
+                lancamento.numero = lancamento.numero + '-' + lancamento.numeroParcela;
                 return _.pick(lancamento, 'idLancamento', 'tipo', 'dataLancamento', 'numero', 'favorecido', 'deposito', 'pagamento', 'saldo', 'historico');
             })
         };
@@ -181,16 +182,30 @@ app.controller('LancamentoController', ['$scope', '$filter', 'LancamentoService'
         };
 
         $scope.excluir = function(conta, idLancamento) {
-            modalExcluir().then(function() {
-                LancamentoService.delete(idLancamento)
-                    .then(function(data) { 
-                        modalMessage("Lançamento Removido com sucesso!");
-                        todos(conta);                        
-                    })
-                    .catch(function(e) {
-                        modalMessage(e);
+            LancamentoService.get(idLancamento)
+                .then(function(lancamento) {
+                    if(lancamento.numeroParcela < lancamento.lancamentoProgramado.numeroParcela) {                
+                        modalMessage("Este lançamento não pode ser excluído. É necessário excluir todos os lançamentos posteriores!");
+                        return;
+                    }     
+                    modalExcluir().then(function() {
+                        LancamentoService.delete(idLancamento)
+                            .then(function(data) { 
+                                modalMessage("Lançamento Removido com sucesso!");
+                                todos(conta);                        
+                            })
+                            .catch(function(e) {
+                                modalMessage(e);
+                            });
                     });
-            });
+                })
+                .catch(function(e) {
+                    modalMessage(e);
+                });
+                
+                
+            
+            
         }; 
         
         var ajustarDados = function(data) {
